Move printuserbooks to UI
